Big refactor of DataStore
Splitting out a couple subsystems: one is a plugin-like system implemented via DataSourceConnector and its subclasses; the other is for managing queued work for DataSourceConnectors.

All screenshotting/pdf rendering stuff is disabled for now. The first of the two above systems has been implemented, but the DataSourceWorkQueue stuff is still in progress.